+++ src/main/java/nextstep/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package nextstep.support.exception;
+package nextstep.exception;
 
 public class BusinessException extends RuntimeException {
