+++ src/main/java/nextstep/web/reservation/ReservationService.java
@@ -1,9 +1,9 @@
 package nextstep.web.reservation;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.support.exception.AuthErrorCode;
-import nextstep.support.exception.BusinessException;
-import nextstep.support.exception.CommonErrorCode;
+import nextstep.exception.AuthErrorCode;
+import nextstep.exception.BusinessException;
+import nextstep.exception.CommonErrorCode;
 import nextstep.web.member.Member;
 import nextstep.web.reservation_waiting.ReservationWaiting;
 import nextstep.web.reservation_waiting.ReservationWaitingDao;
